Remove unneded wget
DragonFly BSD and FreeBSD both come with fetch(1), a utility which is
very similar to wget, so installing wget to pull one file to then
delete it again is entirely unnecessary. This saves a fetch, and a
package deletion, and changes the command invocation to use the
existing fetch command.