Update to latest DragonFly BSD release
In addition to updating to a newer version of DragonFly, the latest as
of this commit, 5.2.1, this changes out the design to use variables in
most places for the ISO name and version so that to update everything
going forward all that needs to be changed is one line in the
Makefile. The Makefile and packer build file were adjusted
accordingly.